feat(add.html, main.js): implement product addition feature with form handling
- Updated add.html to include a Bootstrap-styled form for adding products.
- Enhanced main.js to handle form submission, sending product data to the server via a POST request.
- Cleared initial product data in db.json to start with an empty product list.
- Minor adjustments in index.html for consistency in navigation.

+++ chapter-11/main.js
@@ -1,17 +1,49 @@
 const API = `http://localhost:3001/products`;
+const productList = document.getElementById("productList");
+const productAddForm = document.getElementById("productAddForm");
 
+if (productAddForm) {
+    productAddForm.addEventListener("submit", (e) => {
+        e.preventDefault();
+        addProduct();
+    });
+}
+const addProduct = () => {
+    const product = {
+        name: document.getElementById("name").value,
+        price: document.getElementById("price").value,
+        quantity: document.getElementById("quantity").value,
+        category: document.getElementById("category").value,
+        imageUrl: document.getElementById("image").value,
+    };
+    // call API thêm sản phẩm
+    fetch(`${API}`, {
+        method: "POST",
+        headers: {
+            "Content-Type": "application/json",
+        },
+        body: JSON.stringify(product),
+    })
+        .then(() => {
+            console.log("Thêm sản phẩm thành công");
+            window.location.href = "./index.html";
+        })
+        .catch(() => console.log("Thất bại!"));
+};
 const deleteProduct = (id) => {
+    const confirm = window.confirm("Bạn có chắc chắn muốn xóa sản phẩm này không?");
+    if (!confirm) return;
     fetch(`${API}/${id}`, {
         method: "DELETE",
     })
         .then(() => console.log("Xóa sản phẩm thành công"))
         .catch(() => console.log("Thất bại!"));
 };
-
-const render = () => {
+const renderProduct = () => {
     fetch(API)
         .then((response) => response.json())
         .then((data) => {
+            if (!productList) return;
             productList.innerHTML = data
                 .map(
                     (item, index) => `
@@ -24,9 +56,9 @@ const render = () => {
                     <td>${item.category}</td>
                     <td>
                         <a href="./edit.html?id=${item.id}" class="btn btn-primary">Sửa</a>
-                        <button class="btn btn-danger" onclick="deleteProduct(${
+                        <button class="btn btn-danger" onclick="deleteProduct('${
                             item.id
-                        })">Xóa</button>
+                        }')">Xóa</button>
                     </td>
                 </tr>
             `
@@ -34,4 +66,4 @@ const render = () => {
                 .join("");
         });
 };
-render();
+renderProduct();
